rename: RankCount 패키지 vo로 이동.
- RankCount 클래스가 도메인 역할을 하지 않는 것 같아 변경.
