+++ src/main/java/lotto/domain/Statistics.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import lotto.domain.vo.RankCounts;
 
 public class Statistics {
